+++ src/Game/Text/fr/ScreenText.fr.sjson.lua
@@ -41,7 +41,12 @@ local newData = {
 	-- 	Id = "ModsNikkelMHadesBiomes_Codex_BoonInfo_Orpheus",
 	-- 	DisplayName = "Songs of {$TempTextData.BoonName}",
 	-- },
-	-- #region Relationship hints
+	-- #region Relationship/Unlock hints
+	{
+		Id = "ModsNikkelMHadesBiomes_Codex_Custom_Relationship_Requirement",
+		DisplayName =
+		"\\n {#CodexStandardFormat}{!Icons.CodexEntryIncompleteIcon}  {#CodexChapterClearFormat} Apprenez-en plus en approfondissant votre relation avec {$TempTextData.Name} {#CodexStandardFormat}",
+	},
 	-- {
 	-- 	Id = "ModsNikkelMHadesBiomes_Codex_EurydiceUnlockHint01",
 	-- 	DisplayName =