+++ src/Game/Text/it/ScreenText.it.sjson.lua
@@ -42,7 +42,12 @@ local newData = {
 	-- 	Id = "ModsNikkelMHadesBiomes_Codex_BoonInfo_Orpheus",
 	-- 	DisplayName = "Songs of {$TempTextData.BoonName}",
 	-- },
-	-- #region Relationship hints
+	-- #region Relationship/Unlock hints
+	{
+		Id = "ModsNikkelMHadesBiomes_Codex_Custom_Relationship_Requirement",
+		DisplayName =
+		"\\n {#CodexStandardFormat}{!Icons.CodexEntryIncompleteIcon}  {#CodexChapterClearFormat} Scopri di più approfondendo il tuo rapporto con {$TempTextData.Name} {#CodexStandardFormat}",
+	},
 	-- {
 	-- 	Id = "ModsNikkelMHadesBiomes_Codex_EurydiceUnlockHint01",
 	-- 	DisplayName =