+++ src/Game/Text/ja/ScreenText.ja.sjson.lua
@@ -42,7 +42,12 @@ local newData = {
 	-- 	Id = "ModsNikkelMHadesBiomes_Codex_BoonInfo_Orpheus",
 	-- 	DisplayName = "Songs of {$TempTextData.BoonName}",
 	-- },
-	-- #region Relationship hints
+	-- #region Relationship/Unlock hints
+	{
+		Id = "ModsNikkelMHadesBiomes_Codex_Custom_Relationship_Requirement",
+		DisplayName =
+		"\\n {#CodexStandardFormat}{!Icons.CodexEntryIncompleteIcon}  {#CodexChapterClearFormat} {$TempTextData.Name}との絆が深まると、追記される {#CodexStandardFormat}",
+	},
 	-- {
 	-- 	Id = "ModsNikkelMHadesBiomes_Codex_EurydiceUnlockHint01",
 	-- 	DisplayName =